update App.tsx and hooks/index.ts

+++ hooks/index.ts
@@ -1,9 +1,5 @@
-import * as SecureStore from 'expo-secure-store'
 import {
-  AlertButton,
-  AlertOptions,
   Platform,
-  Alert as DefaultAlert,
 } from 'react-native'
 import { EventRegister } from 'react-native-event-listeners'
 import { MMKV } from 'react-native-mmkv'
@@ -69,16 +65,24 @@ export const sendCustomEvent = (event: string, data?: any) => {
 }
 
 
-export const checkForUpdates = async () => {
+export const checkForUpdates = async (force?: boolean) => {
   try {
     const res = await checkForUpdateAsync()
     if (res.isAvailable) {
       await fetchUpdateAsync()
-      Alert("Update Available", "An update to the app has been downloaded to your device. Click the Update button to install it, alternatively, it will be installed on the next boot of the app",
-        [{ text: 'Dismiss' }, { text: "Update", onPress: async () => await reloadAsync() }]
-      )
+      if (force) {
+        await fetchUpdateAsync()
+        await reloadAsync()
+      }
+      else
+        Alert("Update Available", "An update to the app has been downloaded to your device. Click the Update button to install it, alternatively, it will be installed on the next boot of the app",
+          [{ text: 'Dismiss' }, { text: "Update", onPress: async () => await reloadAsync() }]
+        )
     }
-  } catch { }
+    return true
+  } catch {
+    return false
+  }
 }
 
 export const convertHexToRGBA = (hexCode: string, opacity = 1) => {
